refactor: price parameter
- fix 파라미터 null error
- Integer to Long

+++ src/main/java/com/flab/livecommerce/application/item/command/RegisterItemCommand.java
@@ -11,8 +11,8 @@ public class RegisterItemCommand {
 
     private Long shopId;
     private String name;
-    private Integer price;
-    private Integer salesPrice;
+    private Long price;
+    private Long salesPrice;
     private String description;
     private Integer stockQuantity;
